feat: add metadata to checkout. create component for checkout content
